Add 100-atoi.c file

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -10,7 +10,7 @@
 int _atoi(char *s)
 {
 	int sign = 1;
-	int result = 0;
+	long int result = 0;
 	int num = 0;
 
 	while (*s != '\0')
@@ -30,4 +30,3 @@ int _atoi(char *s)
 	}
 	return (result * sign);
 }
-
